backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2  # Pin to 3.x for passlib 1.7.4 compatibility
python-dotenv==1.0.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Email validation
email-validator==2.1.0

